# swaggers/swagger-concepts.json
{"swagger":"2.0","info":{"description":"\"MedicalConcepts PUJ RESTful API\"","version":"1.1.0","title":"MedicalConcepts PUJ API"},"host":"192.168.56.120:8080","basePath":"/","tags":[{"name":"concept-controller","description":"Concept Controller"}],"paths":{"/concept":{"get":{"tags":["concept-controller"],"summary":"getAllConceptDinamic","operationId":"getAllConceptDinamicUsingGET","produces":["application/json"],"parameters":[{"name":"concept_id","in":"query","description":"concept_id","required":false,"type":"integer","format":"int32"},{"name":"domain_id","in":"query","description":"domain_id","required":false,"type":"string"},{"name":"offset","in":"query","required":false,"type":"integer","format":"int64"},{"name":"pageNumber","in":"query","required":false,"type":"integer","format":"int32"},{"name":"pageSize","in":"query","required":false,"type":"integer","format":"int32"},{"name":"paged","in":"query","required":false,"type":"boolean"},{"name":"short_desc","in":"query","description":"short_desc","required":false,"type":"string"},{"name":"sort.sorted","in":"query","required":false,"type":"boolean"},{"name":"sort.unsorted","in":"query","required":false,"type":"boolean"},{"name":"unpaged","in":"query","required":false,"type":"boolean"},{"name":"vocabulary_id","in":"query","description":"vocabulary_id","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["concept-controller"],"summary":"create","operationId":"createUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"authorization","in":"header","description":"authorization","required":true,"type":"string"},{"in":"body","name":"createConceptDto","description":"createConceptDto","required":true,"schema":{"$ref":"#/definitions/CreateConceptDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["concept-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"authorization","in":"header","description":"authorization","required":true,"type":"string"},{"in":"body","name":"conceptDto","description":"conceptDto","required":true,"schema":{"$ref":"#/definitions/ConceptDto"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/concept/{concept_id}":{"delete":{"tags":["concept-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["application/json"],"parameters":[{"name":"authorization","in":"header","description":"authorization","required":true,"type":"string"},{"name":"concept_id","in":"path","description":"concept_id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"ConceptDto":{"type":"object","required":["anat_or_cond","code","code_billing_status","code_change","code_cms_claim_status","code_planned_type","code_status","codescheme","codetype","codewithperiods","concept_class_id","domain_id","invalid_reason","long_desc","oldpxordx","poa_code_change","poa_code_status","pxordx","sex_cd","short_desc","standard_concept","track","vocabulary_id"],"properties":{"id":{"type":"integer","format":"int32"},"pxordx":{"type":"string"},"oldpxordx":{"type":"string"},"codetype":{"type":"string"},"concept_class_id":{"type":"string"},"concept_id":{"type":"integer","format":"int32"},"vocabulary_id":{"type":"string"},"domain_id":{"type":"string"},"track":{"type":"string"},"standard_concept":{"type":"string"},"code":{"type":"string"},"codewithperiods":{"type":"string"},"codescheme":{"type":"string"},"long_desc":{"type":"string"},"short_desc":{"type":"string"},"code_status":{"type":"string"},"code_change":{"type":"string"},"code_change_year":{"type":"integer","format":"int32"},"code_planned_type":{"type":"string"},"code_billing_status":{"type":"string"},"code_cms_claim_status":{"type":"string"},"sex_cd":{"type":"string"},"anat_or_cond":{"type":"string"},"poa_code_status":{"type":"string"},"poa_code_change":{"type":"string"},"poa_code_change_year":{"type":"integer","format":"int32"},"valid_start_date":{"type":"integer","format":"int32"},"valid_end_date":{"type":"integer","format":"int32"},"invalid_reason":{"type":"string"},"create_dt":{"type":"integer","format":"int32"}},"title":"ConceptDto"},"CreateConceptDto":{"type":"object","required":["anat_or_cond","code","code_billing_status","code_change","code_cms_claim_status","code_planned_type","code_status","codescheme","codetype","codewithperiods","concept_class_id","domain_id","invalid_reason","long_desc","oldpxordx","poa_code_change","poa_code_status","pxordx","sex_cd","short_desc","standard_concept","track","vocabulary_id"],"properties":{"pxordx":{"type":"string"},"oldpxordx":{"type":"string"},"codetype":{"type":"string"},"concept_class_id":{"type":"string"},"concept_id":{"type":"integer","format":"int32"},"vocabulary_id":{"type":"string"},"domain_id":{"type":"string"},"track":{"type":"string"},"standard_concept":{"type":"string"},"code":{"type":"string"},"codewithperiods":{"type":"string"},"codescheme":{"type":"string"},"long_desc":{"type":"string"},"short_desc":{"type":"string"},"code_status":{"type":"string"},"code_change":{"type":"string"},"code_change_year":{"type":"integer","format":"int32"},"code_planned_type":{"type":"string"},"code_billing_status":{"type":"string"},"code_cms_claim_status":{"type":"string"},"sex_cd":{"type":"string"},"anat_or_cond":{"type":"string"},"poa_code_status":{"type":"string"},"poa_code_change":{"type":"string"},"poa_code_change_year":{"type":"integer","format":"int32"},"valid_start_date":{"type":"integer","format":"int32"},"valid_end_date":{"type":"integer","format":"int32"},"invalid_reason":{"type":"string"},"create_dt":{"type":"integer","format":"int32"}},"title":"CreateConceptDto"}}}